requirement.txt
tensorflow==2.13.0
numpy
pandas
opencv-python
matplotlib
scikit-learn
tqdm
streamlit
pillow
scipy